src/test/java/com/danseg/crudusuarios/controller/AuthControllerTest.java
package com.danseg.crudusuarios.controller;

import com.danseg.crudusuarios.model.AuthRequest;
import com.danseg.crudusuarios.model.AuthResponse;
import com.danseg.crudusuarios.util.JwtTokenUtil;
import com.danseg.crudusuarios.service.impl.UserDetailsServiceImpl;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.AuthenticationException;
import org.springframework.security.core.userdetails.UserDetails;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.*;

public class AuthControllerTest {

    @Mock
    private AuthenticationManager authenticationManager;

    @Mock
    private JwtTokenUtil jwtTokenUtil;

    @Mock
    private UserDetailsServiceImpl userDetailsService;

    @InjectMocks
    private AuthController authController;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void testCreateAuthenticationTokenSuccess() throws Exception {
        AuthRequest authRequest = new AuthRequest("user@example.com", "password");
        UserDetails userDetails = mock(UserDetails.class);
        when(userDetails.getUsername()).thenReturn("user@example.com");
        when(userDetailsService.loadUserByUsername("user@example.com")).thenReturn(userDetails);
        when(jwtTokenUtil.generateToken(userDetails)).thenReturn("jwtToken");
        Authentication authentication = mock(Authentication.class);
        when(authenticationManager.authenticate(any(UsernamePasswordAuthenticationToken.class))).thenReturn(authentication);

        ResponseEntity<?> response = authController.createAuthenticationToken(authRequest);

        assertNotNull(response);
        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertTrue(response.getBody() instanceof AuthResponse);
        AuthResponse authResponse = (AuthResponse) response.getBody();
        assertEquals("jwtToken", authResponse.getJwt());
    }

    @Test
    public void testCreateAuthenticationTokenFailure() {
        AuthRequest authRequest = new AuthRequest("user@example.com", "wrongpassword");
        when(authenticationManager.authenticate(any(UsernamePasswordAuthenticationToken.class))).thenThrow(new AuthenticationException("Bad credentials") {});

        Exception exception = assertThrows(Exception.class, () -> {
            authController.createAuthenticationToken(authRequest);
        });

        assertEquals("Credenciales incorrectas + org.springframework.security.core.AuthenticationException: Bad credentials", exception.getMessage());
    }
}
